auth-api/requirements.txt
flask==2.2.5
werkzeug<3.0
flask-cors==3.0.10
pymongo==4.6.1
python-dotenv==1.0.1
flask-jwt-extended==4.6.0
bcrypt==4.1.3